Fix UTF-8 not allowed in Equal field for inhibition rules
This commit fixes a bug where UTF-8 characters are not allowed in
the Equal field for inhibition rules, even when UTF-8 strict mode
is enabled.

This bug occurred because we forgot to override the validation in
model.LabelName. I have copied the same logic used for GroupBy with
GroupByStr, adding EqualStr.

We would like to upgrade prometheus/common in future and use the
validation there instead, but it presents challenges with downstream
projects like Mimir and Cortex where, at present, UTF-8 can be enabled
and disabled in separate components at the same time,
which is not supported in prometheus/common.

The tests for this fix can be found in grafana/prometheus-alertmanager#94
and grafana/alerting#271.